Consume pending input before parsing DSR info.
The `read_dsr` function assumes that there is no pending input on stdin besides
the DSR responses generated by DSR 1337 and DSR 5.

However, if there is any pending input (such as from the user when `it2check` is
run as part of a larger script), `it2check` incorrectly returns false within
`iTerm2.app` and hangs in any terminals that don't respond to DSR 1337 until an
`n` is entered.